+++ tests/test_read.py
@@ -35,7 +35,7 @@ class ReadTest(unittest.TestCase):
 
 				elif i % 4 == 2:
 					self.reads[c][1] = line.strip()
-					
+
 					self.bases['A'] += line.count('A')
 					self.bases['T'] += line.count('T')
 					self.bases['G'] += line.count('G')